PWN-3907 - move TransactionImageView.kt and TransactionSwapImageView.kt to UiKit
