+++ app/build.gradle
@@ -256,7 +256,6 @@ dependencies {
     implementation "io.palaima.debugdrawer:debugdrawer-timber:$debugDrawerVersion"
 
     // Glide
-    implementation 'com.github.bumptech.glide:glide:4.12.0'
     kapt 'com.github.bumptech.glide:compiler:4.12.0'
     implementation 'com.caverock:androidsvg-aar:1.4'
