+++ app/src/main/java/org/p2p/wallet/home/ui/select/SelectTokenViewHolder.kt
@@ -10,7 +10,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.RequestBuilder
 import com.bumptech.glide.request.RequestOptions
-import org.p2p.wallet.common.glide.SvgSoftwareLayerSetter
+import org.p2p.uikit.glide.SvgSoftwareLayerSetter
 import org.p2p.wallet.databinding.ItemTokenSimpleBinding
 import org.p2p.wallet.home.model.Token
 import org.p2p.wallet.utils.dip
