+++ app/src/main/java/org/p2p/wallet/send/ui/dialogs/SendConfirmBottomSheet.kt
@@ -6,10 +6,10 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import org.koin.android.ext.android.inject
+import org.p2p.uikit.glide.GlideManager
 import org.p2p.wallet.R
 import org.p2p.wallet.auth.analytics.AuthAnalytics
 import org.p2p.wallet.auth.interactor.AuthInteractor
-import org.p2p.wallet.common.glide.GlideManager
 import org.p2p.wallet.common.ui.NonDraggableBottomSheetDialogFragment
 import org.p2p.wallet.databinding.DialogSendConfirmBinding
 import org.p2p.wallet.send.analytics.SendAnalytics
