+++ app/src/main/java/org/p2p/wallet/send/ui/main/AccountFeeView.kt
@@ -7,8 +7,8 @@ import androidx.core.view.isInvisible
 import androidx.core.view.isVisible
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
+import org.p2p.uikit.glide.GlideManager
 import org.p2p.wallet.R
-import org.p2p.wallet.common.glide.GlideManager
 import org.p2p.wallet.databinding.WidgetAccountFeeViewBinding
 import org.p2p.wallet.send.model.SendFee
 import org.p2p.wallet.utils.asApproximateUsd
