+++ app/src/main/java/org/p2p/wallet/swap/ui/orca/SwapConfirmBottomSheet.kt
@@ -6,11 +6,11 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import org.koin.android.ext.android.inject
+import org.p2p.uikit.glide.GlideManager
 import org.p2p.wallet.R
 import org.p2p.wallet.auth.interactor.AuthInteractor
-import org.p2p.wallet.common.analytics.interactor.ScreensAnalyticsInteractor
 import org.p2p.wallet.common.analytics.constants.ScreenNames
-import org.p2p.wallet.common.glide.GlideManager
+import org.p2p.wallet.common.analytics.interactor.ScreensAnalyticsInteractor
 import org.p2p.wallet.common.ui.NonDraggableBottomSheetDialogFragment
 import org.p2p.wallet.databinding.DialogSwapConfirmBinding
 import org.p2p.wallet.swap.model.SwapConfirmData
