--- app/src/main/java/org/p2p/wallet/utils/ImageViewExtensions.kt
@@ -1,14 +1,0 @@
-package org.p2p.wallet.utils
-
-import android.widget.ImageView
-import androidx.annotation.DrawableRes
-import androidx.core.view.isVisible
-
-infix fun ImageView.withImageOrGone(@DrawableRes imageRes: Int?) {
-    if (imageRes != null) {
-        isVisible = true
-        setImageResource(imageRes)
-    } else {
-        isVisible = false
-    }
-}
