+++ app/src/main/res/layout/dialog_send_confirm.xml
@@ -51,7 +51,7 @@
         android:textAppearance="@style/WalletTheme.TextAppearance.Medium16"
         app:layout_constraintTop_toBottomOf="@id/titleDivider" />
 
-    <org.p2p.wallet.common.ui.widget.RoundedImageView
+    <org.p2p.uikit.atoms.UiKitRoundedImageView
         android:id="@+id/sourceImageView"
         android:layout_width="44dp"
         android:layout_height="44dp"
@@ -99,7 +99,7 @@
         app:layout_constraintTop_toBottomOf="@id/sourceImageView" />
 
 
-    <org.p2p.wallet.common.ui.widget.RoundedImageView
+    <org.p2p.uikit.atoms.UiKitRoundedImageView
         android:id="@+id/destinationImageView"
         android:layout_width="44dp"
         android:layout_height="44dp"
