+++ app/src/main/res/layout/dialog_transaction_progress.xml
@@ -68,7 +68,7 @@
         app:layout_constraintBottom_toBottomOf="@id/transactionImageView"
         app:layout_constraintTop_toTopOf="@id/transactionImageView" />
 
-    <org.p2p.wallet.common.ui.widget.RoundedImageView
+    <org.p2p.uikit.atoms.UiKitRoundedImageView
         android:id="@+id/transactionImageView"
         android:layout_width="44dp"
         android:layout_height="44dp"
