+++ app/src/main/res/layout/fragment_send.xml
@@ -94,7 +94,7 @@
                             app:layout_constraintEnd_toEndOf="parent"
                             app:layout_constraintTop_toTopOf="@id/availableTextView" />
 
-                        <org.p2p.wallet.common.ui.widget.RoundedImageView
+                        <org.p2p.uikit.atoms.UiKitRoundedImageView
                             android:id="@+id/sourceImageView"
                             android:layout_width="44dp"
                             android:layout_height="44dp"
@@ -211,7 +211,7 @@
                             app:layout_constraintStart_toStartOf="parent"
                             app:layout_constraintTop_toTopOf="parent" />
 
-                        <org.p2p.wallet.common.ui.widget.RoundedImageView
+                        <org.p2p.uikit.atoms.UiKitRoundedImageView
                             android:id="@+id/targetImageView"
                             android:layout_width="44dp"
                             android:layout_height="44dp"
@@ -368,7 +368,7 @@
                         android:layout_width="match_parent"
                         android:layout_height="wrap_content">
 
-                        <org.p2p.wallet.common.ui.widget.RoundedImageView
+                        <org.p2p.uikit.atoms.UiKitRoundedImageView
                             android:id="@+id/networkImageView"
                             android:layout_width="44dp"
                             android:layout_height="44dp"
@@ -437,7 +437,7 @@
                     android:visibility="gone"
                     tools:visibility="visible">
 
-                    <org.p2p.wallet.common.ui.widget.RoundedImageView
+                    <org.p2p.uikit.atoms.UiKitRoundedImageView
                         android:id="@+id/warningImageView"
                         android:layout_width="44dp"
                         android:layout_height="44dp"
