+++ app/src/main/res/layout/fragment_transaction_transfer.xml
@@ -106,7 +106,7 @@
                         android:orientation="vertical"
                         app:layout_constraintGuide_percent="0.5" />
 
-                    <org.p2p.wallet.common.ui.widget.RoundedImageView
+                    <org.p2p.uikit.atoms.UiKitRoundedImageView
                         android:id="@+id/sourceImageView"
                         android:layout_width="44dp"
                         android:layout_height="44dp"
@@ -145,7 +145,7 @@
                         tools:text="~$989.12"
                         tools:visibility="visible" />
 
-                    <org.p2p.wallet.common.ui.widget.RoundedImageView
+                    <org.p2p.uikit.atoms.UiKitRoundedImageView
                         android:id="@+id/transferImageView"
                         android:layout_width="32dp"
                         android:layout_height="32dp"
@@ -161,7 +161,7 @@
                         app:tint="@color/textIconSecondary"
                         tools:ignore="ContentDescription" />
 
-                    <org.p2p.wallet.common.ui.widget.RoundedImageView
+                    <org.p2p.uikit.atoms.UiKitRoundedImageView
                         android:id="@+id/destinationImageView"
                         android:layout_width="44dp"
                         android:layout_height="44dp"
