+++ app/src/main/res/layout/item_token_list.xml
@@ -8,7 +8,7 @@
     android:foreground="?selectableItemBackground"
     android:paddingHorizontal="16dp">
 
-    <org.p2p.wallet.common.ui.widget.RoundedImageView
+    <org.p2p.uikit.atoms.UiKitRoundedImageView
         android:id="@+id/imageView"
         android:layout_width="32dp"
         android:layout_height="32dp"
