+++ app/src/main/res/layout/item_token_simple.xml
@@ -11,7 +11,7 @@
     android:focusable="true"
     android:paddingHorizontal="16dp">
 
-    <org.p2p.wallet.common.ui.widget.RoundedImageView
+    <org.p2p.uikit.atoms.UiKitRoundedImageView
         android:id="@+id/tokenImageView"
         android:layout_width="44dp"
         android:layout_height="44dp"
