+++ app/src/main/res/layout/widget_account_fee_view.xml
@@ -7,7 +7,7 @@
     android:layout_height="wrap_content"
     tools:parentTag="androidx.constraintlayout.widget.ConstraintLayout">
 
-    <org.p2p.wallet.common.ui.widget.RoundedImageView
+    <org.p2p.uikit.atoms.UiKitRoundedImageView
         android:id="@+id/accountImageView"
         android:layout_width="44dp"
         android:layout_height="44dp"
