+++ app/src/main/res/layout/widget_qr_view.xml
@@ -44,7 +44,7 @@
             android:layout_gravity="center"
             android:background="@drawable/bg_circle">
 
-            <org.p2p.wallet.common.ui.widget.RoundedImageView
+            <org.p2p.uikit.atoms.UiKitRoundedImageView
                 android:id="@+id/watermarkImageView"
                 android:layout_width="44dp"
                 android:layout_height="44dp"
