+++ app/src/main/res/values/dimens.xml
@@ -9,8 +9,6 @@
     <dimen name="history_transaction_list_last_item_padding">56dp</dimen>
     <dimen name="history_transaction_item_height">50dp</dimen>
     <dimen name="history_transaction_image_size">50dp</dimen>
-    <dimen name="history_transaction_image_icon_size">18dp</dimen>
-    <dimen name="history_transaction_image_token_icon_size">28dp</dimen>
     <!--TransactionDetails-->
     <dimen name="history_transaction_details_image_size">64dp</dimen>
     <dimen name="history_transaction_details_image_icon_size">24dp</dimen>
