+++ ui-kit/build.gradle
@@ -5,6 +5,9 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.4.1'
     implementation 'androidx.swiperefreshlayout:swiperefreshlayout:1.1.0'
 
+    // Glide
+    api 'com.github.bumptech.glide:glide:4.12.0'
+
     implementation("com.google.android.material:material") {
         version { strictly "1.5.0" }
     }
