+++ ui-kit/src/main/kotlin/org/p2p/uikit/atoms/UiKitTransactionSwapImageView.kt
@@ -1,12 +1,12 @@
-package org.p2p.wallet.common.ui.widget
+package org.p2p.uikit.atoms
 
 import android.content.Context
 import android.util.AttributeSet
 import android.widget.FrameLayout
-import org.p2p.wallet.R
-import org.p2p.wallet.common.glide.GlideManager
-import org.p2p.wallet.databinding.WidgetTransactionSwapImageBinding
-import org.p2p.wallet.utils.viewbinding.inflateViewBinding
+import org.p2p.uikit.R
+import org.p2p.uikit.databinding.WidgetTransactionSwapImageBinding
+import org.p2p.uikit.glide.GlideManager
+import org.p2p.uikit.utils.inflateViewBinding
 
 private const val IMAGE_SIZE = 28
 
@@ -19,11 +19,11 @@ class TransactionSwapImageView @JvmOverloads constructor(
     private val binding = inflateViewBinding<WidgetTransactionSwapImageBinding>()
 
     init {
-        val typedArray = context.obtainStyledAttributes(attrs, R.styleable.TransactionSwapImageView)
+        val typedArray = context.obtainStyledAttributes(attrs, R.styleable.UiKitTransactionSwapImageView)
 
         val tokenIconSize = typedArray.getDimensionPixelSize(
-            R.styleable.TransactionSwapImageView_tokenIconSize,
-            resources.getDimensionPixelSize(R.dimen.history_transaction_image_token_icon_size)
+            R.styleable.UiKitTransactionSwapImageView_tokenIconSize,
+            resources.getDimensionPixelSize(R.dimen.ui_kit_transaction_image_token_icon_size)
         )
 
         with(binding) {
@@ -47,7 +47,7 @@ class TransactionSwapImageView @JvmOverloads constructor(
         }
     }
 
-    private fun RoundedImageView.setViewSize(tokenIconSize: Int) {
+    private fun UiKitRoundedImageView.setViewSize(tokenIconSize: Int) {
         layoutParams = layoutParams.also { params ->
             params.height = tokenIconSize
             params.width = tokenIconSize
