+++ ui-kit/src/main/res/drawable/bg_token_rounded.xml
@@ -4,5 +4,5 @@
     android:shape="rectangle">
     <corners android:radius="16dp" />
 
-    <solid android:color="@color/backgroundAppStroke" />
+    <solid android:color="@color/bg_rain" />
 </shape>
