+++ ui-kit/src/main/res/drawable/ic_state_pending.xml
@@ -5,12 +5,15 @@
     android:viewportWidth="18"
     android:viewportHeight="18">
     <path
-        android:fillColor="@color/systemWarningSecondary"
+        android:fillColor="#5887FF"
         android:pathData="M9,0.5L9,0.5A8.5,8.5 0,0 1,17.5 9L17.5,9A8.5,8.5 0,0 1,9 17.5L9,17.5A8.5,8.5 0,0 1,0.5 9L0.5,9A8.5,8.5 0,0 1,9 0.5z"
         android:strokeWidth="1"
         android:strokeColor="#ffffff" />
     <path
-        android:fillColor="#ffffff"
-        android:fillType="evenOdd"
-        android:pathData="M9.9,4.9C9.9,4.4029 9.497,4 9,4C8.5029,4 8.1,4.4029 8.1,4.9V8.9C8.1,9.3971 8.5029,9.8 9,9.8C9.497,9.8 9.9,9.3971 9.9,8.9V4.9ZM9,12.9C9.5523,12.9 10,12.4523 10,11.9C10,11.3477 9.5523,10.9 9,10.9C8.4477,10.9 8,11.3477 8,11.9C8,12.4523 8.4477,12.9 9,12.9Z" />
+        android:fillColor="#00000000"
+        android:pathData="M9,6V10.3333L12,12"
+        android:strokeWidth="1.8"
+        android:strokeColor="#ffffff"
+        android:strokeLineCap="round"
+        android:strokeLineJoin="round" />
 </vector>
