+++ ui-kit/src/main/res/drawable/ic_transaction_send.xml
@@ -1,0 +1,11 @@
+<vector
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+    <path
+        android:fillColor="@color/black"
+        android:fillType="evenOdd"
+        android:pathData="M11.3064,2.2797C11.6936,1.9068 12.3064,1.9068 12.6936,2.2797L17.6936,7.0945C18.0915,7.4776 18.1034,8.1106 17.7203,8.5084C17.3372,8.9063 16.7042,8.9182 16.3064,8.5351L13,5.3512L13,16C13,16.5523 12.5523,17 12,17C11.4477,17 11,16.5523 11,16L11,5.3512L7.6936,8.5351C7.2958,8.9182 6.6628,8.9063 6.2797,8.5084C5.8966,8.1106 5.9085,7.4776 6.3064,7.0945L11.3064,2.2797ZM4,18C3.4477,18 3,18.4477 3,19C3,19.5523 3.4477,20 4,20H20C20.5523,20 21,19.5523 21,19C21,18.4477 20.5523,18 20,18H4Z" />
+</vector>
