+++ ui-kit/src/main/res/layout/widget_transaction_image.xml
@@ -7,22 +7,23 @@
     android:layout_height="match_parent"
     tools:parentTag="android.widget.FrameLayout">
 
-    <org.p2p.wallet.common.ui.widget.RoundedImageView
+    <org.p2p.uikit.atoms.UiKitRoundedImageView
         android:id="@+id/transactionTokenImageView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
+        android:layout_margin="2dp"
         android:background="@drawable/bg_token_rounded"
         android:padding="12dp"
-        android:layout_margin="2dp"
         android:src="@drawable/ic_transaction_send"
         app:cornerRadius="16dp"
         tools:visibility="visible" />
 
     <ImageView
         android:id="@+id/transactionStatus"
-        android:layout_width="@dimen/history_transaction_image_icon_size"
-        android:layout_height="@dimen/history_transaction_image_icon_size"
+        android:layout_width="@dimen/ui_kit_transaction_image_icon_size"
+        android:layout_height="@dimen/ui_kit_transaction_image_icon_size"
         android:layout_gravity="bottom|end"
+        android:importantForAccessibility="no"
         android:visibility="gone"
         tools:src="@drawable/ic_state_pending"
         tools:visibility="visible" />
