+++ ui-kit/src/main/res/layout/widget_transaction_swap_image.xml
@@ -13,7 +13,7 @@
         android:layout_margin="2dp"
         android:background="@drawable/bg_token_swap_rounded" />
 
-    <org.p2p.wallet.common.ui.widget.RoundedImageView
+    <org.p2p.uikit.atoms.UiKitRoundedImageView
         android:id="@+id/sourceImageView"
         android:layout_width="28dp"
         android:layout_height="28dp"
@@ -22,7 +22,7 @@
         app:cornerRadius="8dp"
         tools:src="@tools:sample/avatars" />
 
-    <org.p2p.wallet.common.ui.widget.RoundedImageView
+    <org.p2p.uikit.atoms.UiKitRoundedImageView
         android:id="@+id/destinationImageView"
         android:layout_width="28dp"
         android:layout_height="28dp"
