+++ ui-kit/src/main/res/values/attrs.xml
@@ -26,6 +26,18 @@
         </attr>
     </declare-styleable>
 
+    <declare-styleable name="RoundedImageView">
+        <attr name="cornerRadius" />
+    </declare-styleable>
+
+    <declare-styleable name="UiKitTransactionImageView">
+        <attr name="statusIconSize" format="dimension" />
+    </declare-styleable>
+
+    <declare-styleable name="UiKitTransactionSwapImageView">
+        <attr name="tokenIconSize" format="dimension" />
+    </declare-styleable>
+
     <!--  Organisms  -->
 
     <declare-styleable name="UiKitButton">
